backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
django-filter>=24.3,<25.0
pymongo>=4.8,<5.0
django-mongodb-backend>=5.1,<6.0
Pillow>=10.4,<11.0
